Replace $timeout with $interval
This commit replaces all instances of `$timeout` with `$interval` in
the notification directive.
An `$interval` with a repeat count of 1
behaves in exactly the same way as `$timeout`, however Angular's
protractor testing framework treats this methods differently.
The full reason for this change is described in https://github.com/jvandemo/angular-growl-notifications/issues/28